pkg/cover/backend: move module discovery into separate file
elf.go is already quite large and messy.
Module discovery is a well separatable chunk of logic,
move it to a dedicated file.
No code changes.
